requirements.txt
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
transformers>=4.36.0
faiss-cpu>=1.7.4
redis>=5.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.0.0
spacy>=3.7.0
prometheus-client>=0.17.0
torch>=2.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6